third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/Element-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on a clone of a declarative shadow tree with shadowrootcustomelementregistry should return the global registry after getting inserted into a document
  assert_equals: expected null but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
